shit i forgot to push last time + weapon stuff
made it so that u can switch elements on head gun with scroll wheel, or by pressing F. when swapping elements, loaded element head will be put back. eg if i have fire bullet loaded, and one extra ammo for it, if i switch element to water, the loaded bullet will be added to the one extra ammo, so now total got 2 fire ammo. also made it so that u can pick up the different guns in any order and wont cause bugs